BigDecimal for monetary values

+++ src/main/java/com/zopa/dev/constants/QuoteConstant.java
@@ -1,10 +1,12 @@
 package com.zopa.dev.constants;
 
+import java.math.BigDecimal;
+
 public class QuoteConstant {
 
     // This could be made configurable
-    public static double INCREMENT_AMOUNT = 100;
-    public static double LOWER_RANGE = 1000;
-    public static double UPPER_RANGE = 15000;
+    public static BigDecimal INCREMENT_AMOUNT = BigDecimal.valueOf(100);
+    public static BigDecimal LOWER_RANGE = BigDecimal.valueOf(1000);
+    public static BigDecimal UPPER_RANGE = BigDecimal.valueOf(15000);
     public static int TERM_MONTHS = 36;
 }
